:truck: chore(api) move target-profile-training-repo to `src/devcomp/`
